celeryproject/requirements.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.1
idna==2.8
kombu==4.2.2.post1
psycopg2==2.7.7
pytz==2018.9
redis==3.0.1
requests==2.21.0
urllib3==1.24.1
vine==1.2.0
